Adds RPC endpoint syncing status monitoring
Adds the ability to check and record if an RPC endpoint is currently syncing.
This information is stored in the database alongside existing RPC measurement data.

Introduces a new 'isSyncing' field to the RPCMeasurement model and database table.
Also adds a new `db-migrate` service which ensures the database is migrated before the other services boot.
This ensures the API depends on the migrations being completed and the database being healthy.

The indexer now fetches the `isSyncing` status from each RPC endpoint, using the `eth_syncing` RPC method.
The API and related types are updated to reflect this new data point.